# investigate adding BMI2 to the haswell kernel