@lemire 2024-09-02
Our CI performance checker reports a performance degradation.