@lemire 2024-09-02
A solution might be to enable BMI2 only under LLVM.